give the diffusion scheduler its own translation unit and test
Model::generate_image and the LCM helpers move to generate_image.cpp
beside the FFI shim, so model.cpp keeps only manifest parsing and the
route. The helpers take DiffusionParams and are declared in engine.h,
which makes the numerics unit-testable: test_diffusion now checks the
4- and 2-step schedules against diffusers and the alphas_cumprod tail
without needing a bundle, so they run when CACTUS_SD_BUNDLE is unset.
The route follows the plan's declared iterative_denoise strategy and
reports which component is missing rather than sniffing component
names, and the guidance embedding no longer divides by zero when
time_cond_proj_dim is 2.
